requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
python-dateutil==2.8.2
